A lot of changes

+++ apps/nodejs.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+# Check if nodejs version 22 is already installed
+if [ "$(node --version | cut -d 'v' -f 2)" == "22" ]; then
+    # nodejs version 22 is already installed
+    return 0
+fi
+
 # install nvm script
 curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.40.1/install.sh | bash
