+++ apps/nvim.sh
@@ -1,6 +1,12 @@
 #!/bin/bash
 
 NVIM_VERSION='0.10.3'
+
+# Check if nvim version 0.10.3 is already installed
+if [ "$(nvim --version | head -n 1 | cut -d ' ' -f 2)" == v"$NVIM_VERSION" ]; then
+    # nvim version 0.10.3 is already installed
+    return 0
+fi
 sudo apt-get install -y \
     cmake \
     curl \
